Add files via upload
So, this is the webpage so far, I'm still working on a log-in, The code for the picture bytes is there and present on both the home and log-in pages. In orger to make the nice GUI you will need the pictures, so just let me know if you want them. Otherwise, just have all three file (Index.html, reset.css, and login.html in the same folder and they will work, if you have them in seperate folders you will need to modify the code.